CRONAPP-3229 - Ajuste Signup
https://cronapp.atlassian.net/browse/CRONAPP-3329

### Problema:
Desenvolvedor não podia escolhar os campos para o Sign Up
### Solução:
Aplicado lógica para o Sign Up ficar com os campos dinamicamente escolhidos pelo desenvolvedor
