+++ project/M/cronapp-rad-project-mobile-cordova/src/main/mobileapp/www/www-autenticacao/i18n/locale_pt_br.json
@@ -133,6 +133,7 @@
   "Signup.view.Back": "Voltar",
   "Signup.view.Signup": "Cadastre-se",
   "UserSuccessfullyRegistered": "Usuário registrado com sucesso!",
+  "UserNotRegistered": "Usuário não registrado!",
   "EmailCanNotBeEmpty": "O e-mail não pode estar vazio",
   "UsernameCanNotBeEmpty": "O usuário não pode estar vazio",
   "PasswordCanNotBeEmpty": "A senha não pode estar vazia",
