+++ project/M/cronapp-rad-project-mobile-cordova/src/main/mobileapp/www/www-autenticacao/js/blockly/auth/Signup.blockly.js
@@ -9,11 +9,17 @@ window.blockly.js.blockly.auth.Signup = window.blockly.js.blockly.auth.Signup ||
  */
 window.blockly.js.blockly.auth.Signup.SignupArgs = [{ description: 'signupUsername', id: '8df902ea' }, { description: 'signupEmail', id: '0c25d725' }, { description: 'signupPassword', id: '9c203fbf' }, { description: 'signupConfirmPassword', id: '067b073b' }];
 window.blockly.js.blockly.auth.Signup.Signup = async function(signupUsername, signupEmail, signupPassword, signupConfirmPassword) {
- var item;
+ var returnJson, httpStatus;
   if (await this.blockly.js.blockly.auth.Signup.isValidSignup(signupUsername, signupEmail, signupPassword, signupConfirmPassword)) {
-    this.cronapi.util.callServerBlocklyAsynchronous('blockly.Signup:signupApp', async function(sender_item) {
-        item = sender_item;
-      this.cronapi.screen.changeView("#/home/login",[  ]);
+    this.cronapi.util.callServerBlocklyAsynchronous('blockly.UserControl:signUp', async function(sender_returnJson) {
+        returnJson = sender_returnJson;
+      httpStatus = this.cronapi.json.getProperty(this.cronapi.json.createObjectFromString(returnJson), 'code');
+      if (httpStatus == '200') {
+        this.cronapi.screen.notify('success',this.cronapi.i18n.translate("UserSuccessfullyRegistered",[  ]));
+        this.cronapi.screen.changeView("#/app/login",[  ]);
+      } else {
+        this.cronapi.screen.notify('warning',this.cronapi.i18n.translate("UserNotRegistered",[  ]));
+      }
     }.bind(this), signupUsername, signupEmail, signupPassword);
   }
 }
@@ -23,7 +29,7 @@ window.blockly.js.blockly.auth.Signup.Signup = async function(signupUsername, si
  */
 window.blockly.js.blockly.auth.Signup.isValidSignupArgs = [{ description: 'signupUsername', id: 'be2325de' }, { description: 'signupEmail', id: 'dc667925' }, { description: 'signupPassword', id: '9a49950f' }, { description: 'signupConfirmPassword', id: 'f6fdc12c' }];
 window.blockly.js.blockly.auth.Signup.isValidSignup = async function(signupUsername, signupEmail, signupPassword, signupConfirmPassword) {
- var item;
+ var returnJson, httpStatus, isValid;
   isValid = true;
   if (this.cronapi.logic.isNullOrEmpty(signupUsername)) {
     this.cronapi.screen.notify('error',this.cronapi.i18n.translate("UsernameCanNotBeEmpty",[  ]));
